// src/types/reports/new-comers-report.d.ts
export interface INewComersReportsInformation {
    employeeId: number;
    employeeNumber: number;
    badgeId: string;
    employeeImageUrl: string;
    fullName: string;
    username: string;
    workEmail: string;
    personalEmail: string;
    workPhone: string;
    mobilePhone: string;
    homePhone: string;
    emergencyPhone: string;
    contactAddress: string;
    birthday: string;
    birthPlace: string;
    joinDate: string;
    grade: string;
    socialBookNo: string;
    idCardNo: string;
    idCardIssueDate: string;
    idCardIssuePlace: string;
    passportNo: string;
    passportIssueDate: string;
    passportExpiryDate: string;
    isLowPerformance: string;
    isBlacklisted: string;
    positionName: string;
    nationalityName: string;
    genderName: string;
    statusName: string;
    statusColor: string;
    companyName: string;
    locationName: string;
    contractTypeName: string;
    educationNames: [],
    maritalStatusName: string;
    buildingName: string;
    dgName: string;
    dcName: string;
    projectName: string;
    projectId: string;
    managerEmail: string;
    resignationStatusName: string;
}